+++ lib/slack-notifier/payload_middleware/at.rb
@@ -6,7 +6,7 @@ module Slack
       class At < Base
         middleware_name :at
 
-        options at: []
+        options :at => []
 
         def call payload={}
           return payload unless payload[:at]
@@ -18,8 +18,7 @@ module Slack
         private
 
           def format_ats ats
-            Array(ats).map { |at| "<#{at_cmd_char(at)}#{at}> " }
-                      .join("")
+            Array(ats).map { |at| "<#{at_cmd_char(at)}#{at}> " }.join("")
           end
 
           def at_cmd_char at