+++ lib/slack-notifier/util/link_formatter.rb
@@ -35,7 +35,7 @@ module Slack
 
         attr_reader :formats
 
-        def initialize string, formats: %i[html markdown]
+        def initialize(string, formats = [:html, :markdown])
           @formats = formats
           @orig    = string.respond_to?(:scrub) ? string.scrub : string
         end