+++ spec/integration/ping_integration_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 # encoding: utf-8
 
-require_relative "../../lib/slack-notifier"
+require(File.expand_path("../../lib/slack-notifier", File.dirname(__FILE__)))
 
 ruby = if defined?(JRUBY_VERSION)
   "jruby #{JRUBY_VERSION}"
@@ -13,4 +13,4 @@ puts "testing with #{ruby}"
 notifier = Slack::Notifier.new ENV["SLACK_WEBHOOK_URL"], username: "notifier"
 notifier.ping "hello", channel: ["#general", "#random"]
 notifier.ping "hello/こんにちは from notifier test script on #{ruby}\225"
-notifier.ping attachments: [{ color: "#1BF5AF", fallback: "fallback", text: "attachment" }]
+notifier.ping :attachments => [{ color: "#1BF5AF", fallback: "fallback", :text => "attachment" }]