+++ spec/lib/slack-notifier/payload_middleware/format_message_spec.rb
@@ -2,20 +2,20 @@
 
 RSpec.describe Slack::Notifier::PayloadMiddleware::FormatMessage do
   it "passes the text through linkformatter with options[:formats]" do
-    subject = described_class.new(:notifier, formats: [:html])
+    subject = described_class.new(:notifier, :formats => [:html])
     expect(Slack::Notifier::Util::LinkFormatter).to receive(:format)
-      .with("hello", formats: [:html])
-    subject.call(text: "hello")
+      .with("hello", :formats => [:html])
+    subject.call(:text => "hello")
 
     subject = described_class.new(:notifier)
     expect(Slack::Notifier::Util::LinkFormatter).to receive(:format)
-      .with("hello", formats: %i[html markdown])
-    subject.call(text: "hello")
+      .with("hello", :formats => [:html, :markdown])
+    subject.call(:text => "hello")
 
-    subject = described_class.new(:notifier, formats: [:markdown])
+    subject = described_class.new(:notifier, :formats => [:markdown])
     expect(Slack::Notifier::Util::LinkFormatter).to receive(:format)
-      .with("hello", formats: [:markdown])
-    subject.call(text: "hello")
+      .with("hello", :formats => [:markdown])
+    subject.call(:text => "hello")
   end
 
   it "returns the payload unmodified if not :text key" do